+++ jdkjson-core/src/main/java/org/jdkstack/jdkjson/core/JsonPathV2.java
@@ -1,5 +1,5 @@
 package org.jdkstack.jdkjson.core;
 
 public class JsonPathV2 {
-    //
+  //
 }
