+++ jdkjson-examples/src/main/java/org/jdkstack/jdkjson/examples/LruV2.java
@@ -1,10 +1,9 @@
 package org.jdkstack.jdkjson.examples;
 
-import org.jdkstack.jdkjson.core.cache.Constants;
-
 import java.util.LinkedHashMap;
 import java.util.Map;
 import java.util.concurrent.locks.ReentrantLock;
+import org.jdkstack.jdkjson.core.cache.Constants;
 
 /**
  * LinkedHashMap实现Lru算法(线程安全).
