+++ jdkjson-jmh/src/main/java/org/jdkstack/jdkjson/jmh/gson/GsonUtil.java
@@ -7,24 +7,24 @@ import com.google.gson.JsonParser;
 
 public class GsonUtil {
 
+  private static Gson gson = new GsonBuilder().create();
+
   private GsonUtil() {
     //
   }
 
-    private static Gson gson = new GsonBuilder().create();
+  public static String bean2Json(Object obj) {
+    return gson.toJson(obj);
+  }
 
-    public static String bean2Json(Object obj) {
-        return gson.toJson(obj);
-    }
+  public static <T> T json2Bean(String jsonStr, Class<T> objClass) {
+    return gson.fromJson(jsonStr, objClass);
+  }
 
-    public static <T> T json2Bean(String jsonStr, Class<T> objClass) {
-        return gson.fromJson(jsonStr, objClass);
-    }
-
-    public static String jsonFormatter(String uglyJsonStr) {
-        Gson gson = new GsonBuilder().setPrettyPrinting().create();
-        JsonParser jp = new JsonParser();
-        JsonElement je = jp.parse(uglyJsonStr);
-        return gson.toJson(je);
-    }
+  public static String jsonFormatter(String uglyJsonStr) {
+    Gson gson = new GsonBuilder().setPrettyPrinting().create();
+    JsonParser jp = new JsonParser();
+    JsonElement je = jp.parse(uglyJsonStr);
+    return gson.toJson(je);
+  }
 }
